T: test auto import improvements
